requirements.txt
fastapi
uvicorn
lyzr
python-dotenv
aenum